pyproject.toml
[tool.poetry]
name = "graduate-work"
version = "0.1.0"
description = ""
authors = ["Ильдар Галеев <ildar-galeev92@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
pydantic-settings = "^2.2.1"
pymongo = "^4.6.3"
motor = "^3.4.0"
pycryptodomex = {version="^3.20.0", markers="sys_platform=='win32'"}
pycryptodome = {version="^3.20.0", markers="sys_platform=='linux'"}
pydantic-mongo = "^2.1.2"
httpx = "^0.27.0"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
pytest-dependency = "^0.6.0"
pytest-cov = "^5.0.0"

[tool.pytest.ini_options]
pythonpath = [
    ".", "src",
]
markers = [
    "repository",
    "services"
]
asyncio_mode="auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
